convex-js-convex-react-query PR 28: fix typing issue (#43126)
I noticed that there were a couple of issues with the types for this library. For example this would not throw a type-error before:

```ts
convexQuery(api.weather.list, { something: 123 });
```

where `something` is not an arg for the `list` query.

It would error at runtime but not at typecheck time.

It now errors as expected

Also I noticed that it was not possible to pass in an empty parameter if there are no required args:

```ts
convexQuery(api.weather.list); // this errors when it should be allowed
```

the `useQuery` function from "convex/react" allows this.

before this PR you were forced to do this:

```ts
convexQuery(api.weather.list, {});
```

This PR fixes that.

I also was unable to `npm install` in this repo due to invalid `workspace:*` dependencies in the `package.json` im not sure why they were there, I have replaced them with actual dependency versions, I hope thats okay?

In addition running `npm run test:types` which runs TSC was failing, I have fixed those issue.

This PR also introduces some vitest type-tests to confirm that all of this is working as expected.

Co-authored-by: Mike Cann <mike.cann@gmail.com>
GitOrigin-RevId: 062925a8813fbfbf1d661c500fd276b0a5494626